i2c task bug fixed

+++ [00]Rednoah3_Haptic_Studio/RedNoah3_App/Core/Src/rednoah3_app.c
@@ -569,7 +569,7 @@ Descript 	: i2c r/w event
 static void i2c_task(void)
 {
     static u8 scope;
-    u8 type, dsize, asize;
+    u16 type, dsize, asize;
 
     if (s_bit.scope_act)
     {
